# eyewear_node_typescript/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Eyewear API"
  },
  "basePath": "/api/v1",
  "parameters": {
    "Authrization": {
      "in": "header",
      "name": "authorization",
      "type": "string",
      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Indhc2VlbUB5b3BtYWlsLmNvbSIsInVzZXJJZCI6IjYxNmE4MmQyNzQwZTk5MzY4OTNmZTdiNCIsImlhdCI6MTYzNzY2NzkxMSwiZXhwIjoxNjQwMDg3MTExfQ.sOiVb2KjS_5mfUJSPKi9Iymx4mr8LJb3UElGuZ3999Q",
      "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Indhc2VlbUB5b3BtYWlsLmNvbSIsInVzZXJJZCI6IjYxNmE4MmQyNzQwZTk5MzY4OTNmZTdiNCIsImlhdCI6MTYzNzY2NzkxMSwiZXhwIjoxNjQwMDg3MTExfQ.sOiVb2KjS_5mfUJSPKi9Iymx4mr8LJb3UElGuZ3999Q"
    },
    "Timezone": {
      "in": "header",
      "name": "timezone",
      "type": "string",
      "example": "Asia/Calcutta",
      "default": "Asia/Calcutta",
      "required": true
    },
    "AcceptLanguage": {
      "in": "header",
      "name": "Accept-Language",
      "type": "string",
      "example": "en/hi",
      "default": "hi",
      "required": true
    }
  },
...
  "paths": {
...
    "/latestproducts": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "get": {
        "summary": "Returns a list of latest products.",
        "description": "get all latest products list.",
        "tags": ["Shop apis"],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/latestproductsSave"
            }
          }
        }
      }
    },
    "/products": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "get all products",
        "description": "get all products API",
        "tags": ["Shop apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "number",
            "example": 1,
            "default": 1
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "array",
              "properties": {
                "price": {
                  "type": "string",
                  "default": null
                },
                "categories": {
                  "type": "string",
                  "default": []
                },
                "rating": {
                  "type": "string",
                  "default": []
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/productsGet"
            }
          }
        }
      }
    },
    "/productdetail": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "get": {
        "summary": "get single product api",
        "description": "get single product API",
        "tags": ["Shop apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "product_id",
            "type": "string",
            "required": true,
            "example": "616325ececd4223283d4f123",
            "default": "616325ececd4223283d4f123"
          },
          {
            "in": "query",
            "name": "user_id",
            "type": "string",
            "example": null,
            "default": null
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/productsGet"
            }
          }
        }
      }
    },
    "/carts": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "get": {
        "summary": "get user cart api",
        "description": "get user cart API",
        "tags": ["Cart apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "type": "string",
            "required": true,
            "example": "616a82d2740e9936893fe7b4",
            "default": "616a82d2740e9936893fe7b4"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/cartGet"
            }
          }
        }
      }
    },
    "/carts/": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "add cart api",
        "description": "add cart API",
        "tags": ["Cart apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "string",
                  "example": "616a82d2740e9936893fe7b4"
                },
                "product_id": {
                  "type": "string",
                  "example": "616325ececd4223283d4f123"
                },
                "quantity": {
                  "required": true,
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/cartAdd"
            }
          }
        }
      }
    },
    "/carts/{id}": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "put": {
        "summary": "update cart api",
        "description": "update cart API",
        "tags": ["Cart apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "example": "617442a33fd5225b3c8f2b5a",
            "default": "617442a33fd5225b3c8f2b5a"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "quantity": {
                  "required": true,
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/cartUpdate"
            }
          }
        }
      }
    },
    "/carts/{id}/": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "delete": {
        "summary": "delete cart api",
        "description": "delete cart API",
        "tags": ["Cart apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "example": "617442a33fd5225b3c8f2b5a",
            "default": "617442a33fd5225b3c8f2b5a"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/cartDelete"
            }
          }
        }
      }
    },
...
    "/uploadimage": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "Profile",
        "description": "upload user image API",
        "tags": ["Profile apis"],
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "in": "formData",
            "name": "photo",
            "type": "file",
            "required": true,
            "description": "Please select photo"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/uploadImage"
            }
          }
        }
      }
    },
    "/changepassword": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "Change Password",
        "description": "Change Password API",
        "tags": ["Profile apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "current_password": {
                  "required": true,
                  "type": "string",
                  "example": "Test@123",
                  "default": "Test@123"
                },
                "new_password": {
                  "required": true,
                  "type": "string",
                  "example": "Test@1234",
                  "default": "Test@1234"
                },
                "confirm_password": {
                  "required": true,
                  "type": "string",
                  "example": "Test@1234",
                  "default": "Test@1234"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/changePassword"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "latestproductsSave": {
      "allOf": [
        {
          "$ref": "#/definitions/latestproductsSave"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "get latest products"
            }
          }
        }
      ]
    },
    "productsGet": {
      "allOf": [
        {
          "$ref": "#/definitions/productsGet"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "get products"
            }
          }
        }
      ]
    },
...
    "cartGet": {
      "allOf": [
        {
          "$ref": "#/definitions/cartGet"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "cart retrived successfullly!"
            }
          }
        }
      ]
    },
    "cartAdd": {
      "allOf": [
        {
          "$ref": "#/definitions/cartAdd"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "add to cart successfullly!"
            }
          }
        }
      ]
    },
    "cartUpdate": {
      "allOf": [
        {
          "$ref": "#/definitions/cartUpdate"
        },
        {
          "type": "object"
        },
        {
          "post": null,
          "properties": {
            "message": {
              "type": "string",
              "example": "update cart successfullly!"
            }
          }
        }
      ]
    },
    "cartDelete": {
      "allOf": [
        {
          "$ref": "#/definitions/cartDelete"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "cart deleted successfullly!"
            }
          }
        }
      ]
    },
...
    "uploadImage": {
      "allOf": [
        {
          "$ref": "#/definitions/uploadImage"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "Data saved successfullly!"
            }
          }
        }
      ]
    },
    "changePassword": {
      "allOf": [
        {
          "$ref": "#/definitions/changePassword"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "Password changed successfullly!"
            }
          }
        }
      ]
    }
  }
}